refactor: usePopper
