+++ lib/CountDown/CountDown.js
@@ -10,6 +10,7 @@ import {
   PopperInfoHeight,
   SVG,
 } from "./styled";
+import { usePopper } from "./usePopper";
 /**
  * @param {import("./types").CountDownProps} props
  */
@@ -27,8 +28,10 @@ export default function CountDown(props) {
   const popperElement = useRef(null);
   const arrowElement = useRef(null);
 
-  const [popperVisible, setPopperVisible] = useState(false);
-  const [popperInstance, setPopperInstance] = useState(null);
+  const { popperVisible, hidePopper, showPopper } = usePopper({
+    refRef: svgElement,
+    popperRef: popperElement,
+  });
 
   const diffHeight = useMemo(() => {
     if (blockHeight > endBlockHeight) {
@@ -48,31 +51,6 @@ export default function CountDown(props) {
 
     return parseInt(v * 100);
   }, [blockHeight, endBlockHeight, startBlockHeight]);
-
-  function showPopper() {
-    setPopperVisible(true);
-    popperInstance?.update?.();
-  }
-
-  function hidePopper() {
-    setPopperVisible(false);
-  }
-
-  useEffect(() => {
-    setPopperInstance(
-      createPopper(svgElement.current, popperElement.current, {
-        placement: "top",
-        modifiers: [
-          {
-            name: "offset",
-            options: {
-              offset: [0, 8],
-            },
-          },
-        ],
-      }),
-    );
-  }, [svgElement, popperElement]);
 
   useEffect(() => {
     const outerRadius = size / 2;
